+++ src/commands/git/stash.ts
@@ -496,7 +496,7 @@ export class StashGitCommand extends QuickCommand<State> {
 			} catch (ex) {
 				Logger.error(ex, context.title);
 
-				const msg: string = ex?.toString() ?? '';
+				const msg: string = ex?.message ?? ex?.toString() ?? '';
 				if (msg.includes('newer version of Git')) {
 					void window.showErrorMessage(`Unable to stash changes. ${msg}`);